Do a single scan pass over every line, skipping opportunistically
I realized that I scan first for the end of the line, and then again
scan the line for the semicolon. The former skips ahead five bytes
because that's the minimum line length, but in practice lines are going
to be longer. That means I was double-reading a bunch of bytes in the
middle.

Instead, I inline the entire line parsing, looking at even fewer bytes.
This is a >10% speedup.